Add a libbpf submodule for use with the selftests
This adds a libbpf submodule under selftest. When the selftest run
script is executed it goes into the libbpf submodule, checks out
a particular supported version git SHA, and rebuilds the libbpf static
archive. The selftests then link against that archive and run.

This also updates the self tests themselves to fix an inconsistency
where they would sometimes timeout if there isn't enough activity on
the machine it's running on (run mmap a bunch of time to trigger the bpf
programs).

Signed-off-by: grantseltzer <grantseltzer@gmail.com>