Fix error message in BPFMap.Update
When trying to obtain an unsafe pointer to the value type, upon failure
the code was incorrectly returning an error message which contained the
type of the key, not the value. This patch fixes it so that the correct
type is returned for the value in the error message.

Additionally this commit contains a small refactor which uses a type
switch rather than if-else statements for readability.