libbpfgo: Improve error message in batch operations
This patch adds support for improved error messages in batch operations.
It seemed as if the return code would be sufficient, but that seems not
to be the case. It turns out that we actually need the errno information
to further figure out what went wrong and communicate that to the
caller.

For example, for batch read (or batch read and delete) operations a
errno value of ENOENT can indicate that the map is empty, and that is
information that should be propogated back to the caller.