Move RWArray helpers into own package
The RWArray is the only part of the helpers package
that was previous used within the libbpfgo package.
This presents an issue when the helpers package wants
to import libbpfgo (creates a circular dependency)

Moving RWArray to a different package solves this issue.

In general, helpers should be able to import libbpfgo but
libbpfgo should not import helpers

Signed-off-by: grantseltzer <grantseltzer@gmail.com>