lazy ksymbols: fix address query with reverse linear search (#283)
In commit 3ea92310, a new KernelSymbolsTable implementation was added
which uses lazy querying and parsing of a stored string copy of the
/proc/kallsyms file.

In order to keep the efficiency of noncached queries, a binary search
was used for querying symbols by address, since the file mostly stores
thevsymbols in order of addresses.

However, on kernels compiled with CONFIG_KALLSYMS_ALL=y, symbols from
other sections may be present in a non-ordered place. This wasn't
taken in consideration in initial implementations, and so a query by
address could have failed despite existing.

These symbols usually appear near the end of the file, so a reverse
linear search is added after a failed binary search.