chore: call bpf_program__attach_iter from go
This changes libbpfgo to call bpf_program__attach_iter directly from
go instead of doing this via a C wrapper. What this means is that
we maintained opts creation in C, but the actual libbpf attach call is
done in go.

This creates two new functions in libbpfgo.h as helpers:
- bpf_iter_attach_opts_new()
- bpf_iter_attach_opts_free()