Fix double-free in `NewModuleFromBufferArgs` (#327)
In
https://github.com/aquasecurity/libbpfgo/commit/0238ec325a967ca48d8958f449e236b5bc432263
the freeing of C strings was changed to use defers. This can cause a
double-free, which in the best case it will produce a crash. The reason
why this happens is that the memory address at `defer` time is captured
for later execution. If `KConfigFilePath` is less than 3, it was being
freed and set to NULL. Once the defer executes on function return, the
same address we already freed will be passed again.

We observed this while upgrading libbpfgo in Parca Agent
(https://github.com/parca-dev/parca-agent/pull/1599).

Signed-off-by: Francisco Javier Honduvilla Coto <javierhonduco@gmail.com>