feat: add missing API to utilize map of maps
In this commit, the libbpfgo.go file undergoes significant changes,
primarily focused on introducing and updating functions related to BPF
maps. The main feature added is the capability to utilize "map of maps",
which involves creating, managing, and interacting with nested maps.

Added Functions:

    Reload
    GetMapByID
    GetBTFFDByID
    SetInnerMap
    MaxEntries (Note: GetMaxEntries is deprecated)

Updated Functions (to deal with maps returned by CreateMap):

    CreateMap
    Name
    Type
    SetType
    KeySize
    ValueSize
    SetValueSize
    Pin
    Unpin
    SetPinPath
    PinPath
    IsPinned
    FileDescriptor
    GetValue
    GetValueFlags
    GetValueReadInto
    Update
    Resize
    GetValueBatch
    GetValueAndDeleteBatch

Miscellaneous:

Aside from the map-related features, this commit also includes various
minor updates and improves error handling, all of which contribute to
the overall robustness of the library.

Co-authored-by: Kemal Akkoyun <kakkoyun@gmail.com>