feat(map): introduce InnerMapInfo()
This allows to query the inner map information from the outer map before
the module is loaded.

Despite acting as a wrapper around bpf_map__inner_map(), which returns a
pointer to a libbpf bpf_map, it is not safe to use it as a BPFMap since
it will never be created. As soon as the module is loaded, the bpf_map
is destroyed. This is the reason why it returns a BPFMapInfo instead.